casting to int in protobuf size args

+++ source/projects/shared/devices/protobuf_decoder.h
@@ -16,12 +16,12 @@ public:
 			delete wrap.proto();
 		}
 		wrap.proto() = new generic_max();
-		static_cast<generic_max*>(wrap.proto())->ParseFromArray(wrap.data(), wrap.size());
+		static_cast<generic_max*>(wrap.proto())->ParseFromArray(wrap.data(), (int) wrap.size());
 	}
 
 	void encode(proto_message_wrapper& wrap) {
 		wrap.vect().reserve(static_cast<generic_max*>(wrap.proto())->ByteSize());
-		static_cast<generic_max*>(wrap.proto())->SerializeToArray(wrap.vect().data(), wrap.vect().size());
+		static_cast<generic_max*>(wrap.proto())->SerializeToArray(wrap.vect().data(), (int) wrap.vect().size());
 	}
 
 private:
